Moves class to its own file - other fixes by CodeClimate
